+++ packages/components/src/components/Checkbox/styles/checkboxStyles.ts
@@ -8,6 +8,7 @@ export const checkboxStyles = sv({
         'tw-peer tw-relative tw-inline-flex tw-bg-surface tw-text-white tw-shrink-0 tw-rounded-medium tw-border tw-border-line-strong group-hover:tw-border-line-strong hover:tw-border-line-strong tw-transition-colors ' +
         'data-[state="checked"]:tw-border-transparent data-[state="indeterminate"]:tw-border-transparent ' +
         'disabled:tw-border-line-strong disabled:tw-bg-surface disabled:tw-cursor-not-allowed data-[state="checked"]:disabled:tw-bg-disabled ' +
+        'data-[readonly="true"]:tw-pointer-events-none ' +
         `${FOCUS_OUTLINE}`,
     variants: {
         size: {