+++ packages/components/src/components/Select/__tests__/Select.ct.tsx
@@ -4,10 +4,10 @@ import { IconIcon } from '@frontify/fondue-icons';
 import { expect, test } from '@playwright/experimental-ct-react';
 import * as sinon from 'sinon';
 
+import { Select } from '../Select';
+
 import { FOCUS_BORDER_CSS, FOCUS_OUTLINE_CSS } from '#/helpers/constants';
 import { MAX_HEIGHT_MARGIN } from '#/utilities/domUtilities';
-
-import { Select } from '../Select';
 
 const SELECT_TEST_ID = 'test-select';
 const SELECT_MENU_TEST_ID = 'fondue-select-menu';
@@ -463,6 +463,7 @@ test('render indicator on selected item', async ({ mount, page }) => {
     const firstItem = page.getByTestId(ITEM_TEST_ID1);
     const chekmarkIcon = firstItem.getByTestId('fondue-icons-check-mark');
     await expect(firstItem).toHaveCSS('background-color', 'rgba(0, 0, 0, 0)');
+    await expect(firstItem).toHaveAttribute('aria-selected', 'false');
     await expect(chekmarkIcon).not.toBeVisible();
 
     await page.keyboard.press('ArrowDown');
@@ -473,6 +474,7 @@ test('render indicator on selected item', async ({ mount, page }) => {
 
     await expect(firstItem).toHaveAttribute('data-selected', 'true');
     await expect(firstItem).toHaveCSS('background-color', 'rgb(234, 234, 228)');
+    await expect(firstItem).toHaveAttribute('aria-selected', 'true');
     await expect(chekmarkIcon).toBeVisible();
 
     await page.keyboard.press('ArrowDown');