+++ packages/fondue/src/components/RadioList/RadioList.stories.tsx
@@ -14,6 +14,11 @@ export default {
     title: 'Legacy Components/RadioList',
     component: RadioList,
     tags: ['autodocs'],
+    parameters: {
+        status: {
+            type: 'deprecated',
+        },
+    },
     argTypes: {
         direction: {
             options: [RadioListDirection.Horizontal, RadioListDirection.Vertical],