+++ packages/tokens/CHANGELOG.md
@@ -30,6 +30,12 @@
 
 - [#2302](https://github.com/Frontify/fondue/pull/2302) [`d010cb5`](https://github.com/Frontify/fondue/commit/d010cb53242c9d9e9112bb369a9af975189ba36a) Thanks [@noahwaldner](https://github.com/noahwaldner)! - feat: overhauled token generation and switched to new token set
 
+## 4.2.2
+
+### Patch Changes
+
+- [#2400](https://github.com/Frontify/fondue/pull/2400) [`8264bdf`](https://github.com/Frontify/fondue/commit/8264bdf1a3321bb941e4704155600905347c36a5) Thanks [@syeo66](https://github.com/syeo66)! - fix(tailwind-token): add tailwind token types
+
 ## 4.2.1
 
 ### Patch Changes